skip all tests on wrong Test::Builder version
Not sure that is the correct way as a user would
not realize the module does not work.

However, I struggle with a chicken-egg situation
as in this case all the typical Test::More mechanisms
don't work because the tests require to load the
module Test::TAPv3 bevore Test::More and by this time
it is already too late.